+++ ui/src/kapacitor/components/HandlerEmpty.js
@@ -5,7 +5,7 @@ const HandlerEmpty = ({onGoToConfig, validationError}) => (
   <div className="endpoint-tab-contents">
     <div className="endpoint-tab--parameters">
       <div className="endpoint-tab--parameters--empty">
-        <p>This handler has not been configured</p>
+        <p>This handler is not enabled</p>
         <div className="form-group form-group-submit col-xs-12 text-center">
           <button
             className="btn btn-primary"